A couple hanger-on classes in the main script
